+++ @generated/graphql/index.ts
@@ -707,6 +707,14 @@ export type FloatQueryOperatorInput = {
   nin: InputMaybe<Array<InputMaybe<Scalars['Float']>>>
 }
 
+/** Person data input to the newsletter. */
+export type IPersonNewsletter = {
+  /** Person's email. */
+  email: Scalars['String']
+  /** Person's name. */
+  name: Scalars['String']
+}
+
 /** Shopping cart input. */
 export type IStoreCart = {
   /** Order information, including `orderNumber` and `acceptedOffer`. */
@@ -858,10 +866,16 @@ export type JsonQueryOperatorInput = {
 }
 
 export type Mutation = {
+  /** Subscribes a new person to the newsletter list. */
+  subscribeToNewsletter: Maybe<PersonNewsletter>
   /** Checks for changes between the cart presented in the UI and the cart stored in the ecommerce platform. If changes are detected, it returns the cart stored on the platform. Otherwise, it returns `null`. */
   validateCart: Maybe<StoreCart>
   /** Updates a web session with the specified values. */
   validateSession: Maybe<StoreSession>
+}
+
+export type MutationSubscribeToNewsletterArgs = {
+  data: IPersonNewsletter
 }
 
 export type MutationValidateCartArgs = {
@@ -900,6 +914,12 @@ export type PageInfo = {
   pageCount: Scalars['Int']
   perPage: Maybe<Scalars['Int']>
   totalCount: Scalars['Int']
+}
+
+/** Newsletter information. */
+export type PersonNewsletter = {
+  /** Person's ID in the newsletter list. */
+  id: Scalars['String']
 }
 
 export type Query = {
@@ -2363,7 +2383,7 @@ export type StoreCollectionMeta = {
   selectedFacets: Array<StoreCollectionFacet>
 }
 
-/** Product collection type. Possible values are `Department`, `Category`, `Brand` or `Cluster`. */
+/** Product collection type. Possible values are `Department`, `Category`, `Brand`, `Cluster`, `SubCategory` or `Collection`. */
 export type StoreCollectionType =
   /** Product brand. */
   | 'Brand'
@@ -2371,8 +2391,12 @@ export type StoreCollectionType =
   | 'Category'
   /** Product cluster. */
   | 'Cluster'
+  /** Product collection. */
+  | 'Collection'
   /** First level of product categorization. */
   | 'Department'
+  /** Third level of product categorization. */
+  | 'SubCategory'
 
 /** Currency information. */
 export type StoreCurrency = {
@@ -2958,6 +2982,14 @@ export type CartItemFragment = {
       valueReference: string
     }>
   }
+}
+
+export type SubscribeToNewsletterMutationVariables = Exact<{
+  data: IPersonNewsletter
+}>
+
+export type SubscribeToNewsletterMutation = {
+  subscribeToNewsletter: { id: string } | null
 }
 
 export type BrowserProductQueryQueryVariables = Exact<{